♻️(web) move ApiLogModel from shared to ingestion
