+++ src/web/infrastructure/repositories/shared/postgres_api_log_repository.py
@@ -1,7 +1,7 @@
 from referentiel.entities.api_log import ApiLog
 from referentiel.repositories.api_log_repository_interface import IApiLogRepository
 
-from infrastructure.django_apps.shared.models.api_log import ApiLogModel
+from infrastructure.django_apps.ingestion.models.api_log import ApiLogModel
 
 
 class PostgresApiLogRepository(IApiLogRepository):
